refac: shop 공고 등록 렌더링 방식 수정
